backend/requirements.txt
Flask
numpy
beam-eye-tracker
openai
Flask-SocketIO
python-dotenv